Inventory Management code and transition code & volumes
Added an inventory manager script that persists throughout scene changes

it will spawn an object named "InventoryManager", with bool values for different items, currently just named "Item#". Once a puzzle is solved, that bool can be turned to true and the item will be in the player's inventory. We can also check whether the room's related item has been acquired to make the room's puzzle activated or not

so long as there is an object named InventoryManager in the first scene, it will not be deleted ever, and you can always call for it using

InventoryManager.GetComponent<s_Inventory>()

I also added scene transition code and volumes, but haven't linked them to any scenes